// src/types/transit/oba.ts
export interface ObaAgency {
  id: string;
  name: string;
  url: string;
  timezone: string;
  lang?: string;
  phone?: string;
  fareUrl?: string;
  email?: string;
}

export interface ObaRoute {
  id: string;
  shortName: string;
  longName?: string;
  description?: string;
  agencyId: string;
  agency?: ObaAgency; // Optional, if full reference is included
  url?: string;
  color?: string;
  textColor?: string;
  type?: number; // OBA route type (e.g., 3 for Bus)
}

export interface CurrentOBARouteDisplayData {
  routeInfo: ObaRoute;
  stops: Place[];
}